+++ tests/src/Opulence/Events/Dispatchers/EventDispatcherTest.php
@@ -9,6 +9,7 @@
 namespace Opulence\Events\Dispatchers;
 
 use Opulence\Events\Event;
+use Opulence\Events\IEvent;
 use Opulence\Tests\Events\Mocks\Listener;
 
 /**
@@ -16,8 +17,10 @@ use Opulence\Tests\Events\Mocks\Listener;
  */
 class EventDispatcherTest extends \PHPUnit\Framework\TestCase
 {
-    /** @var EventDispatcher The dispatcher to use in tests */
+    /** @var SynchronousEventDispatcher The dispatcher to use in tests */
     private $dispatcher = null;
+    /** @var IEvent|\PHPUnit_Framework_MockObject_MockObject The event registry to use in tests */
+    private $eventRegistry = null;
     /** @var Event|\PHPUnit_Framework_MockObject_MockObject The event to use in tests */
     private $event = null;
     /** @var Listener|\PHPUnit_Framework_MockObject_MockObject The mock listener to use in tests */
@@ -28,34 +31,10 @@ class EventDispatcherTest extends \PHPUnit\Framework\TestCase
      */
     public function setUp()
     {
-        $this->dispatcher = new EventDispatcher();
+        $this->eventRegistry = $this->createMock(IEventRegistry::class);
+        $this->dispatcher = new SynchronousEventDispatcher($this->eventRegistry);
         $this->event = $this->getMockForAbstractClass(Event::class);
         $this->listener = $this->createMock(Listener::class);
-    }
-
-    /**
-     * Tests adding listeners
-     */
-    public function testAddingListeners()
-    {
-        $listener1 = [$this->listener, "doNothing1"];
-        $listener2 = [$this->listener, "doNothing2"];
-        $this->dispatcher->registerListener("foo", $listener1);
-        $this->assertEquals([$listener1], $this->dispatcher->getListeners("foo"));
-        $this->assertTrue($this->dispatcher->hasListeners("foo"));
-        $this->dispatcher->registerListener("foo", $listener2);
-        $this->assertEquals([$listener1, $listener2], $this->dispatcher->getListeners("foo"));
-        $this->assertTrue($this->dispatcher->hasListeners("foo"));
-    }
-
-    /**
-     * Tests checking if an event has listeners
-     */
-    public function testCheckingIfEventHasListeners()
-    {
-        $this->assertFalse($this->dispatcher->hasListeners("foo"));
-        $this->dispatcher->registerListener("foo", [$this->listener, "doNothing1"]);
-        $this->assertTrue($this->dispatcher->hasListeners("foo"));
     }
 
     /**
@@ -63,8 +42,13 @@ class EventDispatcherTest extends \PHPUnit\Framework\TestCase
      */
     public function testDispatchingToMultipleListeners()
     {
-        $this->dispatcher->registerListener("foo", [$this->listener, "doNothing1"]);
-        $this->dispatcher->registerListener("foo", [$this->listener, "doNothing2"]);
+        $listeners = [
+            [$this->listener, "doNothing1"],
+            [$this->listener, "doNothing2"]
+        ];
+        $this->eventRegistry->expects($this->once())
+            ->method("getListeners")
+            ->willReturn($listeners);
         $this->listener->expects($this->once())->method("doNothing1")->with($this->event, "foo", $this->dispatcher);
         $this->listener->expects($this->once())->method("doNothing2")->with($this->event, "foo", $this->dispatcher);
         $this->dispatcher->dispatch("foo", $this->event);
@@ -75,7 +59,12 @@ class EventDispatcherTest extends \PHPUnit\Framework\TestCase
      */
     public function testDispatchingToSingleListener()
     {
-        $this->dispatcher->registerListener("foo", [$this->listener, "doNothing1"]);
+        $listeners = [
+            [$this->listener, "doNothing1"]
+        ];
+        $this->eventRegistry->expects($this->once())
+            ->method("getListeners")
+            ->willReturn($listeners);
         $this->listener->expects($this->once())->method("doNothing1")->with($this->event, "foo", $this->dispatcher);
         $this->dispatcher->dispatch("foo", $this->event);
     }
@@ -89,53 +78,17 @@ class EventDispatcherTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Tests getting listeners
-     */
-    public function testGettingListeners()
-    {
-        $this->assertEquals([], $this->dispatcher->getListeners("foo"));
-        $listener1 = [$this->listener, "doNothing1"];
-        $listener2 = [$this->listener, "doNothing2"];
-        $this->dispatcher->registerListener("foo", $listener1);
-        $this->dispatcher->registerListener("foo", $listener2);
-        $this->assertEquals([$listener1, $listener2], $this->dispatcher->getListeners("foo"));
-    }
-
-    /**
-     * Test that a listener cannot be added twice
-     */
-    public function testListenerCannotBeAddedTwice()
-    {
-        $listener = [$this->listener, "doNothing1"];
-        $this->dispatcher->registerListener("foo", $listener);
-        $this->dispatcher->registerListener("foo", $listener);
-        $this->assertEquals([$listener], $this->dispatcher->getListeners("foo"));
-    }
-
-    /**
-     * Tests removing listeners
-     */
-    public function testRemovingListeners()
-    {
-        $listener1 = [$this->listener, "doNothing1"];
-        $listener2 = [$this->listener, "doNothing2"];
-        $this->dispatcher->registerListener("foo", $listener1);
-        $this->dispatcher->registerListener("foo", $listener2);
-        $this->dispatcher->removeListener("foo", $listener2);
-        $this->assertEquals([$listener1], $this->dispatcher->getListeners("foo"));
-        $this->assertTrue($this->dispatcher->hasListeners("foo"));
-        $this->dispatcher->removeListener("foo", $listener1);
-        $this->assertEquals([], $this->dispatcher->getListeners("foo"));
-        $this->assertFalse($this->dispatcher->hasListeners("foo"));
-    }
-
-    /**
      * Tests that the second listener is not called when event propagation is stopped
      */
     public function testSecondListenerIsNotCalledWhenEventPropagationIsStopped()
     {
-        $this->dispatcher->registerListener("foo", [$this->listener, "stopsPropagation"]);
-        $this->dispatcher->registerListener("foo", [$this->listener, "doNothing1"]);
+        $listeners = [
+            [$this->listener, "stopsPropagation"],
+            [$this->listener, "doNothing1"]
+        ];
+        $this->eventRegistry->expects($this->once())
+            ->method("getListeners")
+            ->willReturn($listeners);
         // Need to manually stop the event from propagating in the stub
         $this->listener->expects($this->once())
             ->method("stopsPropagation")